Update stylecomputing.css
i reorganised ur css to make it neater, but i feel like you should try to make broad labels so it is easier to refer and access in future! :D
there is not any unnecessary css on first look, but may come back later after i see your html and the css used.